Added top menus to scrolling lists
To better reflect the actual usability of the top menu,
the buttons now scroll with the list, and disappear as
the user scrolls.

When the user goes "back", the list will return to the
top. This is meant to allow users to go "back" and use
the menu after they have examined a playlist.